+++ src/libs/shouldShowSubscriptionsMenu/index.native.ts
