+++ src/libs/shouldShowSubscriptionsMenu/index.ts
